+++ MarkdownHelper/history.md
@@ -3,6 +3,8 @@
 ## v0.1.4
 
 * "save note to file button" now can handle relative links to files
+* Added collapsible widget capabillity
+  * this adds <details> and <summary> tags
 
 ## v0.1.3